refactor: moved out `body` from `client`
